+++ include/unity/scopes/Scope.h
@@ -142,19 +142,19 @@ public:
     virtual ~Scope();
 
     /**
-    \brief Returns a list of child scopes aggregated by this scope in user-defined order.
-    \return The list of child scopes aggregated by this scope in user-defined order.
+    \brief Returns a list of child scopes aggregated by this scope.
+    \return The list of child scopes aggregated by this scope.
     */
-    virtual ChildScopeList child_scopes_ordered() = 0;
+    virtual ChildScopeList child_scopes() = 0;
 
     /**
-    \brief Sets the ordered list of child scopes aggregated by this scope.
+    \brief Sets the list of child scopes aggregated by this scope.
     \note The only time this call will return false is if the scope cannot write to its config directory.
     This should not happen in real-world usage, but if it does, check the log for more detail.
-    \param child_scopes_ordered The ordered list of child scopes aggregated by this scope.
+    \param child_scopes The list of child scopes aggregated by this scope.
     \return True if the list was successfully set.
     */
-    virtual bool set_child_scopes_ordered(ChildScopeList const& child_scopes_ordered) = 0;
+    virtual bool set_child_scopes(ChildScopeList const& child_scopes) = 0;
 
     /**
     \brief Initiates a search query (overloaded method).