+++ include/unity/scopes/internal/MWScope.h
@@ -59,8 +59,8 @@ public:
                                    VariantMap const& hints,
                                    MWReplyProxy const& reply) = 0;
 
-    virtual ChildScopeList child_scopes_ordered() = 0;
-    virtual bool set_child_scopes_ordered(ChildScopeList const& child_scopes_ordered) = 0;
+    virtual ChildScopeList child_scopes() = 0;
+    virtual bool set_child_scopes(ChildScopeList const& child_scopes) = 0;
 
     virtual bool debug_mode() = 0;