+++ include/unity/scopes/internal/ScopeObjectBase.h
@@ -69,8 +69,8 @@ public:
                                      MWReplyProxy const& reply,
                                      InvokeInfo const& info) = 0;
 
-    virtual ChildScopeList child_scopes_ordered() const = 0;
-    virtual bool set_child_scopes_ordered(ChildScopeList const& child_scopes_ordered) = 0;
+    virtual ChildScopeList child_scopes() const = 0;
+    virtual bool set_child_scopes(ChildScopeList const& child_scopes) = 0;
 
     virtual bool debug_mode() const = 0;
 };