+++ include/unity/scopes/internal/SearchMetadataImpl.h
@@ -55,6 +55,10 @@ public:
     bool has_location() const;
     void remove_location();
 
+    void set_aggregated_keywords(std::set<std::string> const& aggregated_keywords);
+    std::set<std::string> aggregated_keywords() const;
+    bool is_aggregated() const;
+
     virtual VariantMap serialize() const override;
 
     static SearchMetadata create(VariantMap const& var);
@@ -68,6 +72,7 @@ protected:
 private:
     int cardinality_;
     boost::optional<Location> location_;
+    boost::optional<std::set<std::string>> aggregated_keywords_;
 };
 
 } // namespace internal