+++ include/unity/scopes/internal/smartscopes/SSScopeObject.h
@@ -74,8 +74,8 @@ public:
                              MWReplyProxy const& reply,
                              InvokeInfo const& info) override;
 
-    ChildScopeList child_scopes_ordered() const override;
-    bool set_child_scopes_ordered(ChildScopeList const& child_scopes_ordered) override;
+    ChildScopeList child_scopes() const override;
+    bool set_child_scopes(ChildScopeList const& child_scopes) override;
 
     bool debug_mode() const override;