+++ include/unity/scopes/testing/MockScope.h
@@ -38,7 +38,9 @@ namespace testing
 class MockScope : public unity::scopes::Scope, public unity::scopes::testing::MockObject
 {
 public:
-    MockScope() = default;
+    MockScope(std::string const& endpoint = "",
+              std::string const& identity = "")
+        : MockObject(endpoint, identity) {}
 
     MOCK_METHOD3(search, QueryCtrlProxy(std::string const&,
                                         SearchMetadata const&,
@@ -64,9 +66,7 @@ public:
                                          ActionMetadata const&,
                                          PreviewListenerBase::SPtr const&));
 
-    MOCK_METHOD0(child_scopes_ordered, ChildScopeList());
-
-    MOCK_METHOD1(set_child_scopes_ordered, bool(ChildScopeList const&));
+    MOCK_METHOD0(child_scopes, ChildScopeList());
 
     MOCK_METHOD6(search, QueryCtrlProxy(std::string const&,
                                         std::string const&,
@@ -74,6 +74,8 @@ public:
                                         Variant const&,
                                         SearchMetadata const&,
                                         SearchListenerBase::SPtr const&));
+
+    MOCK_METHOD1(set_child_scopes, bool(ChildScopeList const&));
 };
 
 /// @endcond