+++ src/scopes/internal/ScopeBaseImpl.cpp
@@ -181,7 +181,7 @@ void ScopeBaseImpl::set_child_scopes_repo(std::shared_ptr<ChildScopesRepository>
     child_scopes_repo_initialized_ = true;
 }
 
-ChildScopeList ScopeBaseImpl::child_scopes() const
+ChildScopeList ScopeBaseImpl::find_child_scopes() const
 {
     // Get a copy of the registry proxy
     RegistryProxy reg = registry();
@@ -195,30 +195,30 @@ ChildScopeList ScopeBaseImpl::child_scopes() const
     auto all_scopes = reg->list();
     for (auto const& scope : all_scopes)
     {
-        // New scopes are added disabled by default
-        return_list.push_back( ChildScope{scope.first, false} );
+        // New scopes are added enabled by default
+        return_list.push_back( ChildScope{scope.first, scope.second} );
     }
     return return_list;
 }
 
-ChildScopeList ScopeBaseImpl::child_scopes_ordered(ChildScopeList const& child_scopes_unordered) const
+ChildScopeList ScopeBaseImpl::child_scopes(ChildScopeList const& child_scopes_defaulted) const
 {
     lock_guard<mutex> lock(mutex_);
     if (!child_scopes_repo_initialized_)
     {
-        throw LogicException("ScopeBase::child_scopes_ordered() cannot be called from constructor");
+        throw LogicException("ScopeBase::child_scopes() cannot be called from constructor");
     }
-    return child_scopes_repo_ ? child_scopes_repo_->child_scopes_ordered(child_scopes_unordered) : ChildScopeList();
+    return child_scopes_repo_ ? child_scopes_repo_->child_scopes(child_scopes_defaulted) : ChildScopeList();
 }
 
-bool ScopeBaseImpl::set_child_scopes_ordered(ChildScopeList const& child_scopes_ordered)
+bool ScopeBaseImpl::set_child_scopes(ChildScopeList const& child_scopes)
 {
     lock_guard<mutex> lock(mutex_);
     if (!child_scopes_repo_initialized_)
     {
-        throw LogicException("ScopeBase::set_child_scopes_ordered() cannot be called from constructor");
+        throw LogicException("ScopeBase::set_child_scopes() cannot be called from constructor");
     }
-    return child_scopes_repo_ ? child_scopes_repo_->set_child_scopes_ordered(child_scopes_ordered) : false;
+    return child_scopes_repo_ ? child_scopes_repo_->set_child_scopes(child_scopes) : false;
 }
 
 } // namespace internal