+++ src/scopes/internal/SearchMetadataImpl.cpp
@@ -59,6 +59,21 @@ SearchMetadataImpl::SearchMetadataImpl(VariantMap const& var)
     catch (std::exception &e)
     {
     }
+
+    try
+    {
+        it = find_or_throw("SearchMetadataImpl()", var, "aggregated_keywords");
+        std::vector<Variant> keywords_v = it->second.get_array();
+        std::set<std::string> keywords_s;
+        for (auto const& keyword_v : keywords_v)
+        {
+            keywords_s.emplace(keyword_v.get_string());
+        }
+        aggregated_keywords_ = keywords_s;
+    }
+    catch (std::exception &e)
+    {
+    }
 }
 
 void SearchMetadataImpl::set_cardinality(int cardinality)
@@ -96,6 +111,25 @@ void SearchMetadataImpl::remove_location()
     location_ = boost::none;
 }
 
+void SearchMetadataImpl::set_aggregated_keywords(std::set<std::string> const& aggregated_keywords)
+{
+    aggregated_keywords_ = aggregated_keywords;
+}
+
+std::set<std::string> SearchMetadataImpl::aggregated_keywords() const
+{
+    if (is_aggregated())
+    {
+        return *aggregated_keywords_;
+    }
+    return std::set<std::string>();
+}
+
+bool SearchMetadataImpl::is_aggregated() const
+{
+    return bool(aggregated_keywords_);
+}
+
 std::string SearchMetadataImpl::metadata_type() const
 {
     static const std::string t("search_metadata");
@@ -109,6 +143,15 @@ void SearchMetadataImpl::serialize(VariantMap& var) const
     if (location_)
     {
         var["location"] = location_->serialize();
+    }
+    if (aggregated_keywords_)
+    {
+        std::vector<Variant> keywords_v;
+        for (auto const& keyword_s : *aggregated_keywords_)
+        {
+            keywords_v.emplace_back(Variant(keyword_s));
+        }
+        var["aggregated_keywords"] = keywords_v;
     }
 }