+++ RELEASE_NOTES.md
@@ -7,6 +7,9 @@ Changes in version 0.6.15
   - Added _ENABLE_QT_EXPERIMENTAL_ guard to qt headers until that library
     is finalized. Further changes in the unity::scopes::qt namespace are
     expected at this point.
+  - Renamed "child_scopes()" to "find_child_scopes()"
+  - Renamed "child_scopes_ordered()" to "child_scopes()"
+  - Added is_aggregated() and aggregated_keywords() to SearchMetaData
 
 Changes in version 0.6.14
 =========================